player moves with a grabbed balloon
when grabbing a balloon , the player gameobject becomes a child
of the baloon gameobject.  Any movement of the balloon will
result in a movement for the player.

player gameobject's parent becomes root  when you release
a balloon.

A downside of doing it this way is:
Deleting a balloon could accidently delete the player,
so there will have to be checks before a balloon is deleted.